Fix buzzer pulse pattern and threshold int handling
- Buzzer now produces intermittent "tic tic" pattern while alert active
- Threshold inputs use step=1 with int values instead of floats
- Threshold JSON file uses ints for min/max
- Fan stops immediately on recovery (actuators driven every cycle)

+++ src/dashboard/app.py
@@ -147,9 +147,9 @@ async def update_thresholds(request: Request):
 
     for param in thresholds:
         if f"{param}_min" in form:
-            thresholds[param]["min"] = float(form[f"{param}_min"])
+            thresholds[param]["min"] = int(form[f"{param}_min"])
         if f"{param}_max" in form:
-            thresholds[param]["max"] = float(form[f"{param}_max"])
+            thresholds[param]["max"] = int(form[f"{param}_max"])
 
     save_thresholds(thresholds)
